Added test for heatingstatus
