+++ src/HeatKeeper.Server.WebApi.Tests/HeatKeeper.Server.WebApi.Tests.csproj
@@ -9,7 +9,7 @@
       <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
       <PrivateAssets>all</PrivateAssets>
     </PackageReference>
-    <PackageReference Include="CQRS.AspNet.Testing" Version="0.0.4" />
+    <PackageReference Include="CQRS.AspNet.Testing" Version="0.0.5" />
     <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.5.0" />
     <PackageReference Include="xunit" Version="2.4.2" />
     <PackageReference Include="xunit.runner.visualstudio" Version="2.4.5">
