+++ src/HeatKeeper.Server.WebApi.Tests/HeatingStatusTests.cs
@@ -4,6 +4,7 @@ using System.Net.Http;
 using System.Threading.Tasks;
 using CQRS.AspNet.Testing;
 using CQRS.Command.Abstractions;
+using HeatKeeper.Server.Export;
 using HeatKeeper.Server.Measurements;
 using HeatKeeper.Server.Programs;
 using HeatKeeper.Server.Sensors;
@@ -18,84 +19,29 @@ namespace HeatKeeper.Server.WebApi.Tests;
 public class HeatingStatusTests : TestBase
 {
     [Fact]
-    public async Task ShouldSetHeatingStatusOn()
+    public async Task ShouldSetHeatingStatusAccordingToSetPoint()
     {
-        //var setZoneHeatingStatusCommandHandlerMock = Factory;
-
-
-        Mock<ICommandHandler<SetZoneHeatingStatusCommand>> setZoneHeatingStatusCommandHandlerMock = new Mock<ICommandHandler<SetZoneHeatingStatusCommand>>();
-
-
-        var testLocation = await CreateTestLocation();
+        var setZoneHeatingStatusCommandHandlerMock = Factory.MockCommandHandler<SetZoneHeatingStatusCommand>();
+        var exportHeatingStatusToInfluxDbCommandHandlerMock = Factory.MockCommandHandler<ExportHeatingStatusToInfluxDbCommand>();
+        var testLocation = await Factory.CreateTestLocation();
         var janitor = Factory.Services.GetRequiredService<IJanitor>();
+
+        //Note: The setpoint is 20 degrees, so the heating status should be on when the temperature is below 20 degrees minus the hysteresis which is 2 degree
+
         await testLocation.AddLivingRoomMeasurement(10);
         await janitor.Run("SetChannelStates");
-        await testLocation.AddLivingRoomMeasurement(20);
+        setZoneHeatingStatusCommandHandlerMock.VerifyCommandHandler(c => c.HeatingStatus == HeatingStatus.On, Times.Once());
+        exportHeatingStatusToInfluxDbCommandHandlerMock.VerifyCommandHandler(c => c.HeatingStatus == HeatingStatus.On, Times.Once());
+        //Note: The setpoint is 20 degrees, so the heating status should be off when the temperature is above 20 degrees plus the hysteresis which is 2 degree
+
+        await testLocation.AddLivingRoomMeasurement(30);
         await janitor.Run("SetChannelStates");
-
-    }
-
-    private async Task<TestLocation> CreateTestLocation()
-    {
-        var client = Factory.CreateClient();
-
-        var token = await client.AuthenticateAsAdminUser();
-
-        var locationId = await client.CreateLocation(TestData.Locations.Home, token);
-        var livingRoomZoneId = await client.CreateZone(locationId, TestData.Zones.LivingRoom, token);
-
-        await client.CreateMeasurements(TestData.TemperatureMeasurementRequests, token);
-
-        var sensors = await client.GetSensors(livingRoomZoneId, token);
-
-        var livingRoomSensor = sensors.Single(s => s.ExternalId == TestData.Sensors.LivingRoomSensor);
-        await client.UpdateSensor(new UpdateSensorCommand() { SensorId = livingRoomSensor.Id, Name = livingRoomSensor.Name, Description = livingRoomSensor.Description, ZoneId = livingRoomZoneId }, token);
-
-        await client.CreateMeasurements(TestData.TemperatureMeasurementRequests, token);
-
-        var normalProgramId = await client.CreateProgram(new CreateProgramCommand("Normal", locationId), token);
-
-        await client.ActivateProgram(normalProgramId, token);
-
-        var createScheduleCommand = new CreateScheduleCommand(normalProgramId, "DayTime", "0 15,18,21 * * *");
-
-        var scheduleId = await client.CreateSchedule(normalProgramId, createScheduleCommand, token);
-
-        await client.UpdateProgram(new UpdateProgramCommand(normalProgramId, "Away", scheduleId), token);
-
-        var createSetPointCommand = new CreateSetPointCommand(scheduleId, livingRoomZoneId, 20, 2);
-
-        var setPointId = await client.CreateSetPoint(scheduleId, createSetPointCommand, token);
-
-        return new TestLocation(client, token, locationId, livingRoomZoneId, livingRoomSensor.Id, normalProgramId, scheduleId, setPointId);
+        setZoneHeatingStatusCommandHandlerMock.VerifyCommandHandler(c => c.HeatingStatus == HeatingStatus.Off, Times.Once());
+        exportHeatingStatusToInfluxDbCommandHandlerMock.VerifyCommandHandler(c => c.HeatingStatus == HeatingStatus.Off, Times.Once());
     }
 }
 
-public record TestLocation(HttpClient HttpClient, string Token, long LocationId, long LivingRoomZoneId, long LivingRoomSensorId, long NormalProgramId, long ScheduleId, long LivingRoomSetPointId)
-{
-    public async Task AddLivingRoomMeasurement(double temperature)
-    {
-        var measurementCommand = new MeasurementCommand(TestData.Sensors.LivingRoomSensor, MeasurementType.Temperature, RetentionPolicy.Day, temperature, DateTime.UtcNow);
-        await HttpClient.CreateMeasurements(new[] { measurementCommand }, Token);
-    }
-}
-
-public static class MockExtensions
-{
-    public static Mock<T> RegisterMock<T>(this IServiceRegistry registry) where T : class
-    {
-        Mock<T> mock = new Mock<T>();
-        registry.RegisterInstance(mock.Object);
-        return mock;
-    }
-
-    public static Mock<ICommandHandler<T>> Mock<T>(this IServiceRegistry registry)
-    {
-        Mock<ICommandHandler<T>> mock = new Mock<ICommandHandler<T>>();
-        registry.RegisterInstance(mock.Object);
-        return mock;
-    }
 
 
-}
+
